+++ src/example/index.tsx
@@ -1,9 +1,7 @@
-import { create } from "../library/index.ts";
-import todo from "./todo/index.ts";
+import App from "./todo/index.tsx";
 import { StrictMode } from "react";
 import * as ReactDOM from "react-dom/client";
 
-const App = create.app(todo);
 const root = ReactDOM.createRoot(document.body);
 root.render(
   <StrictMode>