+++ src/library/context/index.tsx
@@ -1,5 +1,6 @@
 import { useOptimisedMemo } from "../hooks/index.ts";
-import { AppContext } from "./types.ts";
+import type { AppContext } from "./types.ts";
+import { Props } from "./types.ts";
 import EventEmitter from "eventemitter3";
 import * as React from "react";
 
@@ -11,7 +12,7 @@ export function useApp() {
   return React.useContext(Context);
 }
 
-export default function app(Tree: React.ComponentType): React.ComponentType {
+export function AppContext({ children }: Props): React.ComponentType {
   return () => {
     const context = useOptimisedMemo(
       () => ({
@@ -20,10 +21,6 @@ export default function app(Tree: React.ComponentType): React.ComponentType {
       [],
     );
 
-    return (
-      <Context.Provider value={context}>
-        <Tree />
-      </Context.Provider>
-    );
+    return <Context.Provider value={context}>{children}</Context.Provider>;
   };
 }