+++ src/library/module/renderer/controller/index.ts
@@ -6,8 +6,7 @@ export default function useController<M extends ModuleDefinition>(
   props: Props<M>,
 ) {
   return useOptimisedMemo(() => {
-    const controller = props.options.controller?.(props.actions.controller);
-    if (!controller) return;
+    const controller = props.options.using.actions(props.actions.controller);
     const actions = Object.entries(controller);
     actions.forEach(([name, ƒ]) => props.dispatchers.attach(name, ƒ));
     return controller;