+++ src/library/module/renderer/controller/types.ts
@@ -1,7 +1,7 @@
 import { ModuleDefinition } from "../../../types/index.ts";
+import { UseOptions } from "../../types.ts";
 import { UseActions } from "../actions/types.ts";
 import { UseDispatchers } from "../dispatchers/types.ts";
-import { UseOptions } from "../types.ts";
 import useController from "./index.ts";
 
 export type Props<M extends ModuleDefinition> = {