+++ src/library/module/renderer/dispatchers/types.ts
@@ -1,8 +1,8 @@
-import { UseApp } from "../../../app/types.ts";
+import { UseApp } from "../../../context/types.ts";
 import { ModuleDefinition } from "../../../types/index.ts";
+import { UseOptions } from "../../types.ts";
 import { UseModel } from "../model/types.ts";
 import { UseQueue } from "../queue/types.ts";
-import { UseOptions } from "../types.ts";
 import { UseUpdate } from "../update/types.ts";
 import useDispatchers from "./index.ts";