+++ src/library/module/renderer/lifecycles/types.ts
@@ -1,8 +1,7 @@
 import { ModuleDefinition } from "../../../types/index.ts";
+import { UseOptions } from "../../types.ts";
 import { UseDispatchers } from "../dispatchers/types.ts";
 import { UseElements } from "../elements/types.ts";
-import { UseRouter } from "../router/types.ts";
-import { UseOptions } from "../types.ts";
 import { UseUpdate } from "../update/types.ts";
 import useLifecycles from "./index.ts";
 
@@ -10,7 +9,6 @@ export type Props<M extends ModuleDefinition> = {
   options: UseOptions<M>;
   elements: UseElements;
   dispatchers: UseDispatchers;
-  router: UseRouter;
   update: UseUpdate;
 };