+++ src/library/module/renderer/model/index.ts
@@ -5,7 +5,7 @@ import { Models } from "./utils.ts";
 import * as React from "react";
 
 export default function useModel<M extends ModuleDefinition>(props: Props<M>) {
-  const model = useOptimisedMemo(() => props.options.model ?? {}, []);
+  const model = useOptimisedMemo(() => props.options.using.model ?? {}, []);
 
   return React.useRef<Models<M["Model"]>>(new Models(model, model));
 }