+++ src/library/module/renderer/model/types.ts
@@ -1,6 +1,6 @@
 import useModel from ".";
 import { ModuleDefinition, State } from "../../../types/index.ts";
-import { UseOptions } from "../types.ts";
+import { UseOptions } from "../../types.ts";
 
 export type Props<M extends ModuleDefinition> = {
   options: UseOptions<M>;