+++ src/library/module/renderer/types.ts
@@ -1,15 +1,10 @@
 import { ModuleDefinition } from "../../types/index.ts";
-import { Options } from "../types.ts";
+import { UseOptions } from "../types.ts";
 
 export type ElementName = string;
 
 export type Props<M extends ModuleDefinition> = {
   options: UseOptions<M>;
-};
-
-export type UseOptions<M extends ModuleDefinition> = Options<M> & {
-  name: ElementName;
-  props: M["Props"];
 };
 
 export type Head<T extends any[]> = T extends [infer X, ...any[]] ? X : never;