+++ src/library/module/types.ts
@@ -1,11 +1,12 @@
 import { ControllerDefinition } from "../controller/types.ts";
 import { ModuleDefinition } from "../types/index.ts";
-import { ViewDefinition } from "../view/types.ts";
+import { ViewArgs } from "../view/types.ts";
 
-export type ElementName = string;
-
-export type Options<M extends ModuleDefinition> = {
-  model?: M["Model"];
-  view: ViewDefinition<M>;
-  controller?: ControllerDefinition<M>;
+export type UseOptions<M extends ModuleDefinition> = {
+  using: {
+    model: M["Model"];
+    actions: ControllerDefinition<M>;
+    props?: M["Props"];
+  };
+  children(module: ViewArgs<M>): React.ReactNode;
 };